Add cobra command, config, and flags
